Issue #3312: Move inputs into a separate directory
