+++ src/test/resources/com/puppycrawl/tools/checkstyle/ant/inputs/InputCheckstyleAntTaskWarning.java
@@ -1,4 +1,4 @@
-package com.puppycrawl.tools.checkstyle.ant;
+package com.puppycrawl.tools.checkstyle.ant.inputs;
 
 public final class InputCheckstyleAntTaskWarning {
     int foo;
