+++ src/test/resources/com/puppycrawl/tools/checkstyle/ant/inputs/flawless/InputCheckstyleAntTaskFlawless.java
@@ -1,4 +1,4 @@
-package com.puppycrawl.tools.checkstyle.ant;
+package com.puppycrawl.tools.checkstyle.ant.inputs.flawless;
 
 public final class InputCheckstyleAntTaskFlawless {
     private String foo = "A short line";
